Set columns from geometry to data

Set a column from the geometry context of a spacetime object to the data context.

Usage

set_col(x, .from_geo, .to_data = .from_geo)

set_wts(x, .wt_col = "wt")

set_nbs(x, .nb_col = "nb")

Arguments

x

a spacetime object

.from_geo

the name of the column in the geometry context

.to_data

the name of the new variable to create in the data context

.wt_col

the name of the weights column in the geometry context

.nb_col

the name of neighbor column in the geometry context

Value

A spacetime object with an active data context and a new column from the geometry context.

Details

These functions will reorder the spacetime object to ensure that it is ordered correctly based on the location time columns in the geometry context defined by the loc_col and time_col attributes respectively.

set_wts() and set_nbs() create a new column in the data context with the same name as the column in the geometry context. If a different name is desired use set_col()

Examples

df_fp <- system.file("extdata", "bos-ecometric.csv", package = "sfdep")
geo_fp <- system.file("extdata", "bos-ecometric.geojson", package = "sfdep")

# read in data
df <- readr::read_csv(df_fp, col_types = "ccidD")
geo <- sf::read_sf(geo_fp)

# Create spacetime object called `bos`
bos <- spacetime(df, geo,
                .loc_col = ".region_id",
                .time_col = "time_period")
bos <- activate(bos, "geometry")
#> Error in UseMethod("activate"): no applicable method for 'activate' applied to an object of class "c('spacetime', 'spec_tbl_df', 'tbl_df', 'tbl', 'data.frame')"
bos$nb <- st_contiguity(bos)
#> Error in spdep::poly2nb(x, queen = queen, ...): Not a polygon object
bos$wt <- st_weights(bos$nb)
#> Warning: Unknown or uninitialised column: `nb`.
#> Error in nb2listw(nb, style = style, zero.policy = allow_zero, ...): Not a neighbours list
bos$card <- st_cardinalties(bos$nb)
#> Warning: Unknown or uninitialised column: `nb`.
#> Error: Assigned data `st_cardinalties(bos$nb)` must be compatible with existing data.
#> ✖ Existing data has 1680 rows.
#> ✖ Assigned data has 0 rows.
#> ℹ Only vectors of size 1 are recycled.

set_nbs(bos)
#> Error in set_nbs(bos): `.nb_col` is missing from `geometry`.
set_wts(bos)
#> Error in set_wts(bos): `.wt_col` is missing from `geometry`.
set_col(bos, "card")
#> Error in set_col(bos, "card"): `.from_geo` is missing from `geometry`.
set_col(bos, "card", "cardinalities")
#> Error in set_col(bos, "card", "cardinalities"): `.from_geo` is missing from `geometry`.