Local G*

Local G*

Usage

local_gstar(x, nb, wt, alternative = "two.sided", ...)

local_gstar_perm(x, nb, wt, nsim = 499, alternative = "two.sided", ...)

Arguments

x

A numeric vector.

nb

a neighbor list object for example as created by st_contiguity().

wt

a weights list as created by st_weights().

alternative

default "two.sided". Should be one of "greater", "less", or "two.sided" to specify the alternative hypothesis.

...

methods passed to spdep::localG() or spdep::localG_perm()

nsim

The number of simulations to run.

Value

a data.frame with columns:

gi: the observed statistic

e_gi: the permutation sample mean

var_gi: the permutation sample variance

p_value: the p-value using sample mean and standard deviation

p_folded_sim: p-value based on the implementation of Pysal which always assumes a two-sided test taking the minimum possible p-value

skewness: sample skewness

kurtosis: sample kurtosis

Examples

nb <- st_contiguity(guerry)
wt <- st_weights(nb)
x <- guerry$crime_pers

res <- local_gstar_perm(x, nb, wt)
#> ! attr `self.include` is `TRUE`. Reporting Gi*.
head(res)
#>            gi       e_gi       var_gi    p_value p_sim p_folded_sim   skewness
#> 1 0.014270222 0.01197024 3.527937e-06 0.22075808 0.256        0.128 0.19322074
#> 2 0.015807678 0.01208509 2.671681e-06 0.02275829 0.032        0.016 0.05581188
#> 3 0.015485705 0.01183264 2.783033e-06 0.02854120 0.020        0.010 0.05393347
#> 4 0.008315062 0.01174262 3.648908e-06 0.07275957 0.092        0.046 0.15136961
#> 5 0.009224485 0.01196468 4.768990e-06 0.20955798 0.224        0.112 0.10768431
#> 6 0.008819945 0.01196374 2.219797e-06 0.03485181 0.024        0.012 0.18020054
#>      kurtosis
#> 1 -0.32116082
#> 2  0.07599905
#> 3 -0.28858269
#> 4 -0.01065484
#> 5 -0.39709384
#> 6 -0.08828424

res <- local_gstar(x, nb, wt)
#> ! attr `self.include` is `TRUE`. Reporting Gi*.
head(res)
#> [1]  1.342359  2.595578  2.388872 -1.848187 -1.209736 -2.034142