Compute the Local Bivariate Moran's I Statistic

Given two continuous numeric variables, calculate the bivariate Local Moran's I.

Usage

local_moran_bv(x, y, nb, wt, nsim = 499)

Arguments

x

a numeric vector of same length as y.

y

a numeric vector of same length as x.

nb

a neighbor list object for example as created by st_contiguity().

wt

a weights list as created by st_weights().

nsim

the number of simulations to run.

Value

a data.frame containing two columns Ib and p_sim containing the local bivariate Moran's I and simulated p-values respectively.

Details

The Bivariate Local Moran, like its global counterpart, evaluates the value of x at observation i with its spatial neighbors' value of y. The value of $$I_i^B$$ is just xi * Wyi. Or, in simpler words the local bivariate Moran is the result of multiplying x by the spatial lag of y. Formally it is defined as

\( I_i^B= cx_i\Sigma_j{w_{ij}y_j} \)

References

Local Spatial Autocorrelation (3): Multivariate Local Spatial Autocorrelation, Luc Anselin

See also

Examples

x <- guerry_nb$crime_pers
y <- guerry_nb$wealth
nb <- guerry_nb$nb
wt <- guerry_nb$wt
local_moran_bv(x, y, nb, wt)
#>              Ib p_sim
#> 1   0.093506129 0.410
#> 2  -0.840144689 0.002
#> 3   0.293390275 0.208
#> 4  -0.782969282 0.036
#> 5  -0.343051282 0.030
#> 6  -0.751763780 0.064
#> 7  -1.212286959 0.140
#> 8   0.494768137 0.306
#> 9   0.050875987 0.006
#> 10 -0.033508122 0.428
#> 11  0.119135965 0.406
#> 12 -0.027058053 0.450
#> 13  0.189608296 0.188
#> 14 -0.184931534 0.054
#> 15  0.186274796 0.250
#> 16 -0.042411941 0.296
#> 17  0.006353979 0.434
#> 18 -0.408271216 0.044
#> 19 -0.115344980 0.378
#> 20  0.130059701 0.444
#> 21  1.683274652 0.028
#> 22 -0.003842500 0.496
#> 23 -1.004957821 0.106
#> 24 -1.104136137 0.004
#> 25  0.857180921 0.002
#> 26 -0.169492824 0.014
#> 27  0.834090623 0.246
#> 28 -0.171464343 0.332
#> 29 -0.034233828 0.348
#> 30 -0.051026184 0.262
#> 31 -0.046850788 0.398
#> 32  0.334455413 0.266
#> 33  0.080264717 0.260
#> 34  0.685071940 0.186
#> 35  0.001704856 0.466
#> 36 -0.079714992 0.090
#> 37  0.137434262 0.336
#> 38  0.125025064 0.220
#> 39 -0.068341930 0.172
#> 40  0.504579881 0.086
#> 41 -0.319036396 0.102
#> 42 -0.016497540 0.336
#> 43  0.225196989 0.024
#> 44 -0.224572220 0.354
#> 45 -0.003990683 0.458
#> 46 -1.017640129 0.086
#> 47  0.367618246 0.210
#> 48 -0.550574797 0.220
#> 49  0.441828404 0.028
#> 50 -0.088146734 0.396
#> 51 -0.190742450 0.348
#> 52  0.537482661 0.138
#> 53 -0.039104825 0.408
#> 54  0.305236945 0.098
#> 55 -0.337570608 0.358
#> 56 -0.155974076 0.314
#> 57 -0.526846963 0.164
#> 58 -1.393680416 0.000
#> 59 -0.832978265 0.038
#> 60 -0.326847547 0.152
#> 61 -0.269459435 0.024
#> 62 -0.443421284 0.028
#> 63 -0.158620813 0.408
#> 64 -0.066683215 0.484
#> 65 -0.790102037 0.064
#> 66 -1.406140486 0.136
#> 67 -0.058461082 0.256
#> 68 -0.004884983 0.476
#> 69 -0.125083947 0.376
#> 70 -1.236566942 0.058
#> 71  1.282709560 0.002
#> 72  0.218277696 0.052
#> 73 -0.366089264 0.000
#> 74  1.371117985 0.002
#> 75 -0.075151611 0.212
#> 76 -1.560168225 0.020
#> 77  0.658190968 0.068
#> 78  0.246687028 0.176
#> 79 -0.139720645 0.406
#> 80 -0.206370145 0.282
#> 81 -0.003905568 0.454
#> 82 -0.205045471 0.224
#> 83 -0.534172545 0.000
#> 84 -0.046631364 0.192
#> 85  0.279224938 0.014