Compute the Local Bivariate Moran's I Statistic

Given two continuous numeric variables, calculate the bivariate Local Moran's I.

Usage

local_moran_bv(x, y, nb, wt, nsim = 499)

Arguments

x

a numeric vector of same length as nb.

y

a numeric vector of same length as nb.

nb

a neighbor list object for example as created by st_contiguity().

wt

a weights list as created by st_weights().

nsim

the number of simulations to run.

Details

\( I_i^B= cx_i\Sigma_j{w_{ij}y_j} \)

See also

Examples

x <- guerry_nb$crime_pers
y <- guerry_nb$wealth
nb <- guerry_nb$nb
wt <- guerry_nb$wt
local_moran_bv(x, y, nb, wt)
#>              Ib p_sim
#> 1   0.093506129 0.404
#> 2  -0.840144689 0.004
#> 3   0.293390275 0.200
#> 4  -0.782969282 0.040
#> 5  -0.343051282 0.044
#> 6  -0.751763780 0.062
#> 7  -1.212286959 0.164
#> 8   0.494768137 0.316
#> 9   0.050875987 0.010
#> 10 -0.033508122 0.450
#> 11  0.119135965 0.426
#> 12 -0.027058053 0.470
#> 13  0.189608296 0.140
#> 14 -0.184931534 0.026
#> 15  0.186274796 0.256
#> 16 -0.042411941 0.334
#> 17  0.006353979 0.472
#> 18 -0.408271216 0.040
#> 19 -0.115344980 0.386
#> 20  0.130059701 0.410
#> 21  1.683274652 0.022
#> 22 -0.003842500 0.500
#> 23 -1.004957821 0.152
#> 24 -1.104136137 0.000
#> 25  0.857180921 0.002
#> 26 -0.169492824 0.012
#> 27  0.834090623 0.180
#> 28 -0.171464343 0.334
#> 29 -0.034233828 0.348
#> 30 -0.051026184 0.248
#> 31 -0.046850788 0.394
#> 32  0.334455413 0.276
#> 33  0.080264717 0.242
#> 34  0.685071940 0.160
#> 35  0.001704856 0.488
#> 36 -0.079714992 0.102
#> 37  0.137434262 0.362
#> 38  0.125025064 0.212
#> 39 -0.068341930 0.152
#> 40  0.504579881 0.076
#> 41 -0.319036396 0.084
#> 42 -0.016497540 0.334
#> 43  0.225196989 0.020
#> 44 -0.224572220 0.422
#> 45 -0.003990683 0.486
#> 46 -1.017640129 0.094
#> 47  0.367618246 0.266
#> 48 -0.550574797 0.236
#> 49  0.441828404 0.044
#> 50 -0.088146734 0.390
#> 51 -0.190742450 0.376
#> 52  0.537482661 0.128
#> 53 -0.039104825 0.456
#> 54  0.305236945 0.096
#> 55 -0.337570608 0.334
#> 56 -0.155974076 0.310
#> 57 -0.526846963 0.174
#> 58 -1.393680416 0.000
#> 59 -0.832978265 0.034
#> 60 -0.326847547 0.132
#> 61 -0.269459435 0.030
#> 62 -0.443421284 0.032
#> 63 -0.158620813 0.402
#> 64 -0.066683215 0.454
#> 65 -0.790102037 0.070
#> 66 -1.406140486 0.150
#> 67 -0.058461082 0.244
#> 68 -0.004884983 0.482
#> 69 -0.125083947 0.414
#> 70 -1.236566942 0.042
#> 71  1.282709560 0.002
#> 72  0.218277696 0.034
#> 73 -0.366089264 0.000
#> 74  1.371117985 0.002
#> 75 -0.075151611 0.222
#> 76 -1.560168225 0.026
#> 77  0.658190968 0.066
#> 78  0.246687028 0.190
#> 79 -0.139720645 0.400
#> 80 -0.206370145 0.264
#> 81 -0.003905568 0.496
#> 82 -0.205045471 0.210
#> 83 -0.534172545 0.000
#> 84 -0.046631364 0.234
#> 85  0.279224938 0.006