Calculate spatial weights

Calculate polygon spatial weights from a nb list. See spdep::nb2listw() for further details.

Usage

st_weights(nb, style = "W", allow_zero = NULL, ...)

Arguments

nb

A neighbor list object as created by st_neighbors().

style

Default "W" for row standardized weights. This value can also be "B", "C", "U", "minmax", and "S". See spdep::nb2listw() for details.

Details

Under the hood, st_weights() creates a listw object and then extracts the weights elements from it as the neighbours element is already--presumably--already existent in the neighbors list you've already created. listw objects are recreated using recreate_listw() when calculating other statistics.

See also

Examples

guerry %>%
 dplyr::mutate(nb = st_contiguity(geometry),
               wt = st_weights(nb))
#> Simple feature collection with 85 features and 28 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 47680 ymin: 1703258 xmax: 1031401 ymax: 2677441
#> CRS:           NA
#> # A tibble: 85 × 29
#>    code_dept count ave_id_geo  dept region department   crime_pers crime_prop
#>  * <fct>     <dbl>      <dbl> <int> <fct>  <fct>             <int>      <int>
#>  1 01            1         49     1 E      Ain               28870      15890
#>  2 02            1        812     2 N      Aisne             26226       5521
#>  3 03            1       1418     3 C      Allier            26747       7925
#>  4 04            1       1603     4 E      Basses-Alpes      12935       7289
#>  5 05            1       1802     5 E      Hautes-Alpes      17488       8174
#>  6 07            1       2249     7 S      Ardeche            9474      10263
#>  7 08            1      35395     8 N      Ardennes          35203       8847
#>  8 09            1       2526     9 S      Ariege             6173       9597
#>  9 10            1      34410    10 E      Aube              19602       4086
#> 10 11            1       2807    11 S      Aude              15647      10431
#> # … with 75 more rows, and 21 more variables: literacy <int>, donations <int>,
#> #   infants <int>, suicides <int>, main_city <ord>, wealth <int>,
#> #   commerce <int>, clergy <int>, crime_parents <int>, infanticide <int>,
#> #   donation_clergy <int>, lottery <int>, desertion <int>, instruction <int>,
#> #   prostitutes <int>, distance <dbl>, area <int>, pop1831 <dbl>,
#> #   geometry <MULTIPOLYGON>, nb <nb>, wt <list>

# using geometry column directly
nb <- st_contiguity(guerry$geometry)
wt <- st_weights(nb)